Get basic component support in place
SML tags now return lazy-evaluated components with Reduce functions, to ensure we only build their trees when necessary. The addition of the Context interface (which is passed to all Reduce calls), allows for passing of arbitrary state downward into the tree. This is intended for app-global state without the need to pass it everywhere else.